Isolated login logic into its own utils file
